requirements.txt
numpy>=1.24
matplotlib>=3.7
Pillow>=10.0
ipympl>=0.9
